+++ LogicBuilder.EntityFrameworkCore.CosmosDb.Tests/Data/BaseDataClass.cs
@@ -1,0 +1,8 @@
+﻿using LogicBuilder.Data;
+
+namespace LogicBuilder.EntityFrameworkCore.CosmosDb.Tests.Data
+{
+    abstract public class BaseDataClass : BaseData
+    {
+    }
+}